+++ fuel-tx/src/tests/offset.rs
@@ -5,6 +5,7 @@
 
 use crate::{
     field::{
+        BlobId as BlobIdField,
         InputContract,
         Inputs,
         MintAmount,
@@ -70,6 +71,14 @@ struct TestedFields {
 fn chargeable_transaction_parts<Tx>(tx: &Tx, bytes: &[u8], cases: &mut TestedFields)
 where
     Tx: Buildable,
+{
+    inputs_assert(tx, bytes, cases);
+    outputs_assert(tx, bytes, cases);
+}
+
+fn inputs_assert<Tx>(tx: &Tx, bytes: &[u8], cases: &mut TestedFields)
+where
+    Tx: Inputs + Buildable,
 {
     tx.inputs().iter().enumerate().for_each(|(idx, i)| {
         let input_ofs = tx
@@ -247,8 +256,6 @@ where
             }
         }
     });
-
-    outputs_assert(tx, bytes, cases);
 }
 
 fn outputs_assert<Tx: Outputs>(tx: &Tx, bytes: &[u8], cases: &mut TestedFields) {
@@ -503,6 +510,51 @@ fn tx_offset_upload() {
         .take(number_cases)
         .for_each(|(tx, _)| {
             let bytes = tx.to_bytes();
+            chargeable_transaction_parts(&tx, &bytes, &mut cases);
+        });
+
+    // Chargeable parts
+    assert!(cases.utxo_id);
+    assert!(cases.owner);
+    assert!(cases.asset_id);
+    assert!(cases.predicate_coin);
+    assert!(cases.predicate_message);
+    assert!(cases.predicate_data_coin);
+    assert!(cases.predicate_data_message);
+    assert!(cases.contract_balance_root);
+    assert!(cases.contract_state_root);
+    assert!(cases.contract_id);
+    assert!(cases.sender);
+    assert!(cases.recipient);
+    assert!(cases.message_data);
+    assert!(cases.message_predicate);
+    assert!(cases.message_predicate_data);
+    assert!(cases.output_to);
+    assert!(cases.output_asset_id);
+    assert!(cases.output_balance_root);
+    assert!(cases.output_contract_state_root);
+    assert!(cases.output_contract_created_state_root);
+    assert!(cases.output_contract_created_id);
+}
+
+#[test]
+fn tx_offset_blob() {
+    let mut cases = TestedFields::default();
+    let number_cases = 100;
+
+    // The seed will define how the transaction factory will generate a new transaction.
+    // Different seeds might implicate on how many of the cases we cover - since we
+    // assert coverage for all scenarios with the boolean variables above, we need to
+    // pick a seed that, with low number of cases, will cover everything.
+    TransactionFactory::<_, Blob>::from_seed(1295)
+        .take(number_cases)
+        .for_each(|(tx, _)| {
+            let bytes = tx.to_bytes();
+
+            // Blob id
+            let offs = tx.blob_id_offset();
+            assert_eq!(bytes[offs..offs + BlobId::LEN], **tx.blob_id());
+
             chargeable_transaction_parts(&tx, &bytes, &mut cases);
         });