+++ fuel-vm/src/tests/test_helpers.rs
@@ -55,10 +55,19 @@ pub fn run_script(script: Vec<Instruction>) -> Vec<Receipt> {
 }
 
 /// Assert that transaction didn't panic
+#[track_caller]
 pub fn assert_success(receipts: &[Receipt]) {
     if let Receipt::ScriptResult { result, .. } = receipts.last().unwrap() {
         if *result != ScriptExecutionResult::Success {
-            panic!("Expected vm success, got {result:?} instead");
+            let Some(Receipt::Panic { reason, .. }) = receipts.get(receipts.len() - 2)
+            else {
+                panic!("Expected vm success, got {result:?} instead (panic receipt missing!)");
+            };
+
+            panic!(
+                "Expected vm success, got {result:?} ({:?}) instead",
+                reason.reason()
+            );
         }
     } else {
         unreachable!("No script result");
@@ -66,6 +75,7 @@ pub fn assert_success(receipts: &[Receipt]) {
 }
 
 /// Assert that transaction receipts end in a panic with the given reason
+#[track_caller]
 pub fn assert_panics(receipts: &[Receipt], reason: PanicReason) {
     if let Receipt::ScriptResult { result, .. } = receipts.last().unwrap() {
         if *result != ScriptExecutionResult::Panic {
@@ -85,5 +95,64 @@ pub fn assert_panics(receipts: &[Receipt], reason: PanicReason) {
         );
     } else {
         unreachable!("No script receipt for a paniced tx");
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum RunResult<T> {
+    Success(T),
+    UnableToExtractValue,
+    Revert,
+    Panic(PanicReason),
+    GenericFailure(u64),
+}
+
+impl<T> RunResult<T> {
+    pub fn is_ok(&self) -> bool {
+        matches!(self, RunResult::Success(_))
+    }
+
+    pub fn map<F: FnOnce(T) -> R, R>(self, f: F) -> RunResult<R> {
+        match self {
+            RunResult::Success(v) => RunResult::Success(f(v)),
+            RunResult::UnableToExtractValue => RunResult::UnableToExtractValue,
+            RunResult::Revert => RunResult::Revert,
+            RunResult::Panic(r) => RunResult::Panic(r),
+            RunResult::GenericFailure(v) => RunResult::GenericFailure(v),
+        }
+    }
+
+    /// Extract the value from the receipts, using the provided extractor function
+    /// to get even more data about successfull runs.
+    pub fn extract(
+        receipts: &[Receipt],
+        value_extractor: fn(&[Receipt]) -> Option<T>,
+    ) -> RunResult<T> {
+        let Receipt::ScriptResult { result, .. } = receipts.last().unwrap() else {
+            unreachable!("No script result");
+        };
+
+        match *result {
+            ScriptExecutionResult::Success => match value_extractor(receipts) {
+                Some(v) => RunResult::Success(v),
+                None => RunResult::UnableToExtractValue,
+            },
+            ScriptExecutionResult::Revert => RunResult::Revert,
+            ScriptExecutionResult::Panic => RunResult::Panic({
+                let Receipt::Panic { reason, .. } = receipts[receipts.len() - 2] else {
+                    unreachable!("No panic receipt");
+                };
+                *reason.reason()
+            }),
+            ScriptExecutionResult::GenericFailure(value) => {
+                RunResult::GenericFailure(value)
+            }
+        }
+    }
+}
+
+impl RunResult<()> {
+    pub fn extract_novalue(receipts: &[Receipt]) -> RunResult<()> {
+        Self::extract(receipts, |_| Some(()))
     }
 }