+++ fuel-vm/src/util.rs
@@ -65,11 +65,14 @@ macro_rules! script_with_data_offset {
                     fuel_types::bytes::padded_len,
                     prelude::Immediate18,
                 };
-                ($tx_offset
+                let value: Immediate18 = $tx_offset
                     .saturating_add(Script::script_offset_static())
                     .saturating_add(
                         padded_len(script_bytes.as_slice()).unwrap_or(usize::MAX),
-                    ) as Immediate18)
+                    )
+                    .try_into()
+                    .expect("script data offset is too large");
+                value
             }
         };
         // re-evaluate and return the finalized script with the correct data offset length
@@ -129,6 +132,8 @@ pub mod test_helpers {
             Outputs,
             ReceiptsRoot,
         },
+        BlobBody,
+        BlobIdExt,
         ConsensusParameters,
         Contract,
         ContractParameters,
@@ -155,6 +160,7 @@ pub mod test_helpers {
         },
         Address,
         AssetId,
+        BlobId,
         BlockHeight,
         ChainId,
         ContractId,
@@ -322,6 +328,12 @@ pub mod test_helpers {
             self
         }
 
+        pub fn with_free_gas_costs(&mut self) -> &mut TestBuilder {
+            let gas_costs = GasCosts::free();
+            self.consensus_params.set_gas_costs(gas_costs);
+            self
+        }
+
         pub fn base_asset_id(&mut self, base_asset_id: AssetId) -> &mut TestBuilder {
             self.consensus_params.set_base_asset_id(base_asset_id);
             self
@@ -424,11 +436,7 @@ pub mod test_helpers {
             initial_balance: Option<(AssetId, Word)>,
             initial_state: Option<Vec<StorageSlot>>,
         ) -> CreatedContract {
-            let storage_slots = if let Some(slots) = initial_state {
-                slots
-            } else {
-                Default::default()
-            };
+            let storage_slots = initial_state.unwrap_or_default();
 
             let salt: Salt = self.rng.gen();
             let program: Witness = contract
@@ -467,6 +475,33 @@ pub mod test_helpers {
                 contract_id,
                 salt,
             }
+        }
+
+        pub fn setup_blob(&mut self, data: Vec<u8>) {
+            let id = BlobId::compute(data.as_slice());
+
+            let tx = TransactionBuilder::blob(BlobBody {
+                id,
+                witness_index: 0,
+            })
+            .add_witness(data.into())
+            .max_fee_limit(self.max_fee_limit)
+            .maturity(Default::default())
+            .add_random_fee_input()
+            .finalize()
+            .into_checked(self.block_height, &self.consensus_params)
+            .expect("failed to check tx");
+
+            let interpreter_params =
+                InterpreterParams::new(self.gas_price, &self.consensus_params);
+            let mut transactor = Transactor::<_, _, _>::new(
+                MemoryInstance::new(),
+                self.storage.clone(),
+                interpreter_params,
+            );
+
+            self.execute_tx_inner(&mut transactor, tx)
+                .expect("Expected vm execution to be successful");
         }
 
         fn execute_tx_inner<M, Tx, Ecal>(